+++ src/main/java/com/example/groutine/domain/challenge/dto/request/ChallengeMissionRequestDto.java
@@ -1,6 +1,6 @@
 package com.example.groutine.domain.challenge.dto.request;
 
-import com.example.groutine.domain.participation.entity.ChallengeMissionType;
+import com.example.groutine.domain.mission.entity.ChallengeMissionType;
 
 public record ChallengeMissionRequestDto(
     String title,
