+++ src/main/java/com/example/groutine/domain/challenge/dto/response/ChallengeDetailResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.groutine.domain.challenge.dto.response;
 
-import com.example.groutine.domain.participation.entity.ChallengeMissionType;
+import com.example.groutine.domain.mission.entity.ChallengeMissionType;
 import lombok.Builder;
 import lombok.Getter;
