+++ src/main/java/com/example/groutine/domain/challenge/service/ChallengeCommandService.java
@@ -4,13 +4,13 @@ import com.example.groutine.domain.challenge.dto.request.ChallengeRequestDto;
 import com.example.groutine.domain.challenge.dto.response.ChallengeIdResponseDto;
 import com.example.groutine.domain.challenge.entity.Challenge;
 import com.example.groutine.domain.challenge.entity.ChallengeMember;
-import com.example.groutine.domain.participation.entity.ChallengeMission;
+import com.example.groutine.domain.mission.entity.ChallengeMission;
 import com.example.groutine.domain.challenge.mapper.ChallengeMapper;
 import com.example.groutine.domain.challenge.mapper.ChallengeMissionMapper;
 import com.example.groutine.domain.challenge.repository.ChallengeRepository;
 import com.example.groutine.domain.member.entity.Member;
-import com.example.groutine.domain.participation.repository.ChallengeMemberRepository;
-import com.example.groutine.domain.participation.status.ChallengeParticipationErrorStatus;
+import com.example.groutine.domain.mission.repository.ChallengeMemberRepository;
+import com.example.groutine.domain.mission.status.ChallengeParticipationErrorStatus;
 import com.example.groutine.global.common.exception.RestApiException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
