+++ src/main/java/com/example/groutine/domain/member/entity/Member.java
@@ -2,7 +2,7 @@ package com.example.groutine.domain.member.entity;
 
 import com.example.groutine.domain.challenge.entity.Challenge;
 import com.example.groutine.domain.challenge.entity.ChallengeMember;
-import com.example.groutine.domain.participation.entity.ChallengeMissionVerification;
+import com.example.groutine.domain.mission.entity.ChallengeMissionVerification;
 import com.example.groutine.global.common.base.BaseEntity;
 import com.example.groutine.global.common.enums.Status;
 import jakarta.persistence.*;
