+++ src/main/java/com/example/groutine/domain/mission/repository/ChallengeMissionVerificationRepository.java
@@ -1,6 +1,6 @@
-package com.example.groutine.domain.participation.repository;
+package com.example.groutine.domain.mission.repository;
 
-import com.example.groutine.domain.participation.entity.ChallengeMissionVerification;
+import com.example.groutine.domain.mission.entity.ChallengeMissionVerification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChallengeMissionVerificationRepository extends JpaRepository<ChallengeMissionVerification, Long>{
