+++ src/main/java/com/example/groutine/domain/mission/service/ChallengeParticipationService.java
@@ -1,0 +1,12 @@
+package com.example.groutine.domain.mission.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class ChallengeParticipationService {
+
+}
